Refactor: paper controller & service 수정
