+++ .eslintrc.json
@@ -15,5 +15,5 @@
   "parserOptions": {
     "project": "./tsconfig.json"
   },
-  "ignorePatterns": ["node_modules/", "dist/", "models/"]
+  "ignorePatterns": ["node_modules/", "dist/", "models/", "app.js", "server.js"]
 }
